backend/requirements.txt
fastapi
uvicorn
tensorflow
numpy
pillow
requests
cryptography
python-multipart

#db

pymongo[srv]
motor
python-dotenv
pydantic[email]